Move point cloud loader file into seperate folder.
